refactor(engine): Reorganize dsl module
