+++ tests/integration/test_workflows.py
@@ -20,8 +20,8 @@ from pathlib import Path
 import pytest
 from temporalio.common import RetryPolicy
 
+from tracecat.dsl.common import DSLInput
 from tracecat.dsl.dispatcher import dispatch_workflow
-from tracecat.dsl.workflow import DSLInput
 
 DATA_PATH = Path(__file__).parent.parent.joinpath("data/workflows")
 SHARED_TEST_DEFNS = list(DATA_PATH.glob("shared_*.yml"))
