+++ tracecat/dsl/workflow.py
@@ -5,10 +5,8 @@ from collections import defaultdict
 from collections.abc import Callable, Coroutine
 from datetime import timedelta
 from enum import StrEnum, auto
-from pathlib import Path
-from typing import Annotated, Any, Literal, Self, TypedDict
+from typing import Any, TypedDict
 
-import yaml
 from temporalio import activity, workflow
 
 from tracecat.contexts import RunContext
@@ -16,103 +14,21 @@ from tracecat.contexts import RunContext
 with workflow.unsafe.imports_passed_through():
     import jsonpath_ng.lexer  # noqa
     import jsonpath_ng.parser  # noqa
+    from pydantic import BaseModel
+
     from tracecat.auth.credentials import Role
     from tracecat.auth.sandbox import AuthSandbox
-    from tracecat.logging import logger
-    from pydantic import BaseModel, ConfigDict, Field, model_validator
-
-    from tracecat.auth import Role
-    from tracecat.contexts import ctx_role, ctx_run, ctx_logger
-    from tracecat.registry import registry
     from tracecat import templates
-
-
-SLUG_PATTERN = r"^[a-z0-9_]+$"
-ACTION_TYPE_PATTERN = r"^[a-z0-9_.]+$"
-
-
-class DSLError(ValueError):
-    pass
-
-
-class Trigger(BaseModel):
-    type: Literal["schedule", "webhook"]
-    ref: str = Field(pattern=SLUG_PATTERN)
-    args: dict[str, Any] = Field(default_factory=dict)
-
-
-class DSLConfig(BaseModel):
-    scheduler: Literal["static", "dynamic"] = "dynamic"
-
-
-class DSLInput(BaseModel):
-    """DSL definition for a workflow.
-
-    The difference between this and a normal workflow engine is that here,
-    our workflow execution order is defined by the DSL itself, independent
-    of a workflow scheduler.
-
-    With a traditional
-    This allows the execution of the workflow to be fully deterministic.
-    """
-
-    model_config = ConfigDict(arbitrary_types_allowed=True)
-    title: str
-    description: str
-    entrypoint: str
-    actions: list[ActionStatement]
-    config: DSLConfig = Field(default_factory=DSLConfig)
-    triggers: list[Trigger] = Field(default_factory=list)
-    inputs: dict[str, Any] = Field(default_factory=dict)
-
-    @staticmethod
-    def from_yaml(path: str | Path) -> DSLInput:
-        with Path(path).open("r") as f:
-            yaml_str = f.read()
-        dsl_dict = yaml.safe_load(yaml_str)
-        return DSLInput.model_validate(dsl_dict)
-
-    def to_yaml(self, path: str | Path) -> None:
-        with Path(path).expanduser().resolve().open("w") as f:
-            yaml.dump(self.model_dump(), f)
-
-    def dump_yaml(self) -> str:
-        return yaml.dump(self.model_dump())
-
-    @model_validator(mode="after")
-    def validate_input(self) -> Self:
-        if not self.actions:
-            raise DSLError("At least one action must be defined")
-        if len({action.ref for action in self.actions}) != len(self.actions):
-            raise DSLError("All task.ref must be unique")
-        valid_actions = tuple(action.ref for action in self.actions)
-        if self.entrypoint not in valid_actions:
-            raise DSLError(f"Entrypoint must be one of the actions {valid_actions!r}")
-        n_entrypoints = sum(1 for action in self.actions if not action.depends_on)
-        if n_entrypoints != 1:
-            raise DSLError(f"Expected 1 entrypoint, got {n_entrypoints}")
-        return self
+    from tracecat.contexts import ctx_logger, ctx_role, ctx_run
+    from tracecat.dsl.common import ActionStatement, DSLInput
+    from tracecat.logging import logger
+    from tracecat.registry import registry
+    from tracecat.db.schemas import Secret  # noqa
 
 
 class DSLRunArgs(BaseModel):
     role: Role
     dsl: DSLInput
-
-
-class ActionStatement(BaseModel):
-    ref: str = Field(pattern=SLUG_PATTERN)
-    """Unique reference for the task"""
-
-    action: str = Field(pattern=ACTION_TYPE_PATTERN)
-    """Namespaced action type"""
-
-    args: dict[str, Any] = Field(default_factory=dict)
-    """Arguments for the action"""
-
-    depends_on: list[str] = Field(default_factory=list)
-    """Task dependencies"""
-
-    run_if: Annotated[str | None, Field(default=None), templates.TemplateValidator()]
 
 
 class DSLContext(TypedDict):
@@ -291,24 +207,14 @@ class DSLWorkflow:
                 return
 
             self.logger.info("Executing task")
-            # TODO: Securely inject secrets into the task arguments
-            # 1. Find all secrets in the task arguments
-            # 2. Load the secrets
-            # 3. Inject the secrets into the task arguments using an enriched context
-
-            # Resolve all templated arguments
-            processed_args = templates.eval_templated_object(
-                task.args, operand=self.context
-            )
-
             # TODO: Set a retry policy for the activity
             activity_result = await workflow.execute_activity(
                 "run_udf",
-                arg=UDFActionArgs(
-                    type=task.action,
-                    args=processed_args,
+                arg=UDFActionInput(
+                    task=task,
                     role=self.role,
-                    context=self.run_ctx,
+                    run_context=self.run_ctx,
+                    exec_context=self.context,
                 ),
                 start_to_close_timeout=timedelta(minutes=1),
             )
@@ -334,11 +240,11 @@ class DSLWorkflow:
         self.scheduler.mark_task(task_ref, marker)
 
 
-class UDFActionArgs(BaseModel):
-    type: str
-    args: dict[str, Any]
+class UDFActionInput(BaseModel):
+    task: ActionStatement
     role: Role
-    context: RunContext
+    exec_context: dict[str, Any]
+    run_context: RunContext
 
 
 class DSLActivities:
@@ -347,23 +253,38 @@ class DSLActivities:
 
     @staticmethod
     @activity.defn
-    async def run_udf(action: UDFActionArgs) -> Any:
-        ctx_run.set(action.context)
-        ctx_role.set(action.role)
-        act_logger = logger.bind(wf_id=action.context.wf_id, role=action.role)
+    async def run_udf(input: UDFActionInput) -> Any:
+        ctx_run.set(input.run_context)
+        ctx_role.set(input.role)
+        task = input.task
+
+        act_logger = logger.bind(
+            task_ref=task.ref, wf_id=input.run_context.wf_id, role=input.role
+        )
+
+        # Securely inject secrets into the task arguments
+        # 1. Find all secrets in the task arguments
+        # 2. Load the secrets
+        # 3. Inject the secrets into the task arguments using an enriched context
+        secret_refs = templates.extract_templated_secrets(task.args)
+        logger.warning("Secrets", secret_refs=secret_refs)
+        async with AuthSandbox(secrets=secret_refs, target="context") as sandbox:
+            # Resolve all templated arguments
+            logger.info("Evaluating task arguments", secrets=sandbox.secrets)
+            args = templates.eval_templated_object(
+                task.args, operand={**input.exec_context, "SECRETS": sandbox.secrets}
+            )
+        type = task.action
         ctx_logger.set(act_logger)
 
-        udf = registry[action.type]
+        udf = registry[type]
         act_logger.info(
-            "Run udf",
-            type=action.type,
-            is_async=udf.is_async,
-            args=action.args,
+            "Run udf", task_ref=task.ref, type=type, is_async=udf.is_async, args=args
         )
         if udf.is_async:
-            result = await udf.fn(**action.args)
+            result = await udf.fn(**args)
         else:
-            result = await asyncio.to_thread(udf.fn, **action.args)
+            result = await asyncio.to_thread(udf.fn, **args)
         act_logger.info("Result", result=result)
         return result
