+++ tracecat/templates/__init__.py
@@ -1,5 +1,10 @@
-from .eval import eval_templated_object
+from .eval import eval_templated_object, extract_templated_secrets
 from .expressions import TemplateExpression
 from .validators import TemplateValidator
 
-__all__ = ["TemplateValidator", "TemplateExpression", "eval_templated_object"]
+__all__ = [
+    "TemplateValidator",
+    "TemplateExpression",
+    "eval_templated_object",
+    "extract_templated_secrets",
+]
